Add in a max update variable
To stop from overloading the backend, put in a click counter to only update the backend every timer click if there are changes to the machine status.

Otherwise, the default is to wait 10x longer than the ping time on the network.

So, if the ping is 30secs to ping local network, then it will update the backend every 300secs (5 minutes) unless there are changes.